Use Starsector's bundled JRE
I don't know if this is a good idea, but I thought I'd suggest it so I could find out why it's this way around :)

If we specify the JRE bundled with starsector then the debug environment in the IDE will more closely match the run environment when not using the IDE (at least on the developer's machine). I don't know if Starsector ships with a modified copy of the JRE, but I know that lots of users use JRE 8 instead, which might be relevant.